+++ lib/ectomancer/oban_bridge.ex
@@ -127,7 +127,7 @@ if Code.ensure_loaded?(Oban) do
           unquote(oban_authorize_call(auth_handler))
 
           handle(fn _params, _actor ->
-            Ectomancer.ObanBridge.list_queues()
+            Ectomancer.ObanBridge.Queries.list_queues()
           end)
         end
       end
@@ -145,7 +145,7 @@ if Code.ensure_loaded?(Oban) do
 
           handle(fn params, _actor ->
             queue_name = params["queue_name"] || params[:queue_name]
-            Ectomancer.ObanBridge.get_queue_depth(queue_name)
+            Ectomancer.ObanBridge.Queries.get_queue_depth(queue_name)
           end)
         end
       end
@@ -175,7 +175,7 @@ if Code.ensure_loaded?(Oban) do
               |> Enum.reject(fn {_k, v} -> is_nil(v) end)
               |> Enum.into(%{})
 
-            Ectomancer.ObanBridge.list_stuck_jobs(filters)
+            Ectomancer.ObanBridge.Queries.list_stuck_jobs(filters)
           end)
         end
       end
@@ -193,7 +193,7 @@ if Code.ensure_loaded?(Oban) do
 
           handle(fn params, _actor ->
             job_id = params["job_id"] || params[:job_id]
-            Ectomancer.ObanBridge.retry_job(job_id)
+            Ectomancer.ObanBridge.Queries.retry_job(job_id)
           end)
         end
       end
@@ -211,249 +211,46 @@ if Code.ensure_loaded?(Oban) do
 
           handle(fn params, _actor ->
             job_id = params["job_id"] || params[:job_id]
-            Ectomancer.ObanBridge.cancel_job(job_id)
+            Ectomancer.ObanBridge.Queries.cancel_job(job_id)
           end)
         end
       end
     end
 
-    # Public API implementations
-
     @doc false
-    def list_queues do
-      import Ecto.Query
-
-      # Check repo is configured before attempting query
-      _ = repo()
-
-      try do
-        queues =
-          Oban.Job
-          |> select(
-            [j],
-            {j.queue, fragment("COUNT(*)"),
-             fragment("SUM(CASE WHEN state = 'executing' THEN 1 ELSE 0 END)"),
-             fragment("SUM(CASE WHEN state = 'available' THEN 1 ELSE 0 END)"),
-             fragment("SUM(CASE WHEN state = 'retryable' THEN 1 ELSE 0 END)"),
-             fragment("SUM(CASE WHEN state = 'discarded' THEN 1 ELSE 0 END)")}
-          )
-          |> group_by([j], j.queue)
-          |> order_by([j], j.queue)
-          |> repo_all()
-
-        formatted_queues =
-          Enum.map(queues, fn {queue, total, executing, available, retryable, discarded} ->
-            %{
-              queue: queue,
-              total: total || 0,
-              executing: executing || 0,
-              available: available || 0,
-              retryable: retryable || 0,
-              discarded: discarded || 0
-            }
-          end)
-
-        {:ok, %{queues: formatted_queues}}
-      rescue
-        e ->
-          {:error, "Failed to list queues: #{Exception.message(e)}"}
-      end
-    end
-
+    defdelegate list_queues(), to: Ectomancer.ObanBridge.Queries
     @doc false
-    def get_queue_depth(queue_name) when is_binary(queue_name) do
-      import Ecto.Query
-
-      # Check repo is configured before attempting query
-      _ = repo()
-
-      try do
-        counts =
-          Oban.Job
-          |> where([j], j.queue == ^queue_name)
-          |> select([j], {
-            fragment("COUNT(*)"),
-            fragment("SUM(CASE WHEN state = 'executing' THEN 1 ELSE 0 END)"),
-            fragment("SUM(CASE WHEN state = 'available' THEN 1 ELSE 0 END)"),
-            fragment("SUM(CASE WHEN state = 'retryable' THEN 1 ELSE 0 END)"),
-            fragment("SUM(CASE WHEN state = 'discarded' THEN 1 ELSE 0 END)")
-          })
-          |> repo_one()
-
-        case counts do
-          nil ->
-            {:ok,
-             %{
-               queue: queue_name,
-               total: 0,
-               executing: 0,
-               available: 0,
-               retryable: 0,
-               discarded: 0
-             }}
-
-          {total, executing, available, retryable, discarded} ->
-            {:ok,
-             %{
-               queue: queue_name,
-               total: total || 0,
-               executing: executing || 0,
-               available: available || 0,
-               retryable: retryable || 0,
-               discarded: discarded || 0
-             }}
-        end
-      rescue
-        e ->
-          {:error, "Failed to get queue depth: #{Exception.message(e)}"}
-      end
-    end
-
-    def get_queue_depth(_), do: {:error, "queue_name must be a string"}
-
+    def get_queue_depth(queue_name) when is_binary(queue_name),
+      do: Ectomancer.ObanBridge.Queries.get_queue_depth(queue_name)
     @doc false
-    def list_stuck_jobs(filters \\ %{}) do
-      import Ecto.Query
-
-      # Check repo is configured before attempting query
-      _ = repo()
-
-      try do
-        query =
-          Oban.Job
-          |> where([j], j.state == "executing")
-
-        # Apply optional filters
-        query =
-          Enum.reduce(filters, query, fn
-            {:queue, queue}, q ->
-              where(q, [j], j.queue == ^queue)
-
-            {:worker, worker}, q ->
-              where(q, [j], j.worker == ^worker)
-
-            {:min_age_minutes, minutes}, q ->
-              cutoff = DateTime.utc_now() |> DateTime.add(-minutes, :minute)
-              where(q, [j], j.attempted_at < ^cutoff)
-
-            _, q ->
-              q
-          end)
-
-        query =
-          query
-          |> order_by([j], desc: j.attempted_at)
-          |> limit(^Map.get(filters, :limit, 100))
-
-        jobs = repo_all(query)
-
-        formatted_jobs =
-          Enum.map(jobs, fn job ->
-            %{
-              id: job.id,
-              queue: job.queue,
-              worker: job.worker,
-              args: job.args,
-              state: job.state,
-              attempt: job.attempt,
-              max_attempts: job.max_attempts,
-              attempted_at: job.attempted_at,
-              inserted_at: job.inserted_at,
-              errors: job.errors
-            }
-          end)
-
-        {:ok, %{jobs: formatted_jobs, count: length(formatted_jobs)}}
-      rescue
-        e ->
-          {:error, "Failed to list stuck jobs: #{Exception.message(e)}"}
-      end
-    end
-
+    def get_queue_depth(other),
+      do: Ectomancer.ObanBridge.Queries.get_queue_depth(other)
     @doc false
-    def retry_job(job_id) when is_integer(job_id) do
-      import Ecto.Query
-
-      # Check repo is configured before attempting query
-      _ = repo()
-
-      try do
-        # Update the job to be available again
-        {count, _} =
-          Oban.Job
-          |> where([j], j.id == ^job_id)
-          |> where([j], j.state in ["retryable", "discarded"])
-          |> repo_update_all(set: [state: "available", scheduled_at: DateTime.utc_now()])
-
-        if count > 0 do
-          {:ok, %{message: "Job #{job_id} scheduled for retry", job_id: job_id}}
-        else
-          {:error, "Job #{job_id} not found or not in retryable/discarded state"}
-        end
-      rescue
-        e ->
-          {:error, "Failed to retry job: #{Exception.message(e)}"}
-      end
-    end
-
-    def retry_job(_), do: {:error, "job_id must be an integer"}
-
+    def list_stuck_jobs(), do: Ectomancer.ObanBridge.Queries.list_stuck_jobs()
     @doc false
-    def cancel_job(job_id) when is_integer(job_id) do
-      import Ecto.Query
-
-      # Check repo is configured before attempting query
-      _ = repo()
-
-      try do
-        # Cancel the job if it's not already completed
-        {count, _} =
-          Oban.Job
-          |> where([j], j.id == ^job_id)
-          |> where([j], j.state != "completed")
-          |> repo_delete_all()
-
-        if count > 0 do
-          {:ok, %{message: "Job #{job_id} cancelled", job_id: job_id}}
-        else
-          {:error, "Job #{job_id} not found or already completed"}
-        end
-      rescue
-        e ->
-          {:error, "Failed to cancel job: #{Exception.message(e)}"}
-      end
-    end
-
-    def cancel_job(_), do: {:error, "job_id must be an integer"}
-
-    # Helper functions to work with configured repo
-
+    defdelegate list_stuck_jobs(filters), to: Ectomancer.ObanBridge.Queries
     @doc false
-    def repo do
-      Application.get_env(:ectomancer, :repo) ||
-        raise ArgumentError,
-              "Ectomancer repo not configured. Set config :ectomancer, :repo, YourApp.Repo"
-    end
-
+    def retry_job(job_id) when is_integer(job_id),
+      do: Ectomancer.ObanBridge.Queries.retry_job(job_id)
     @doc false
-    def repo_all(query) do
-      repo().all(query)
-    end
-
+    def retry_job(other),
+      do: Ectomancer.ObanBridge.Queries.retry_job(other)
     @doc false
-    def repo_one(query) do
-      repo().one(query)
-    end
-
+    def cancel_job(job_id) when is_integer(job_id),
+      do: Ectomancer.ObanBridge.Queries.cancel_job(job_id)
     @doc false
-    def repo_update_all(query, opts) do
-      repo().update_all(query, opts)
-    end
-
+    def cancel_job(other),
+      do: Ectomancer.ObanBridge.Queries.cancel_job(other)
     @doc false
-    def repo_delete_all(query) do
-      repo().delete_all(query)
-    end
+    defdelegate repo(), to: Ectomancer.ObanBridge.Queries
+    @doc false
+    defdelegate repo_all(query), to: Ectomancer.ObanBridge.Queries
+    @doc false
+    defdelegate repo_one(query), to: Ectomancer.ObanBridge.Queries
+    @doc false
+    defdelegate repo_update_all(query, opts), to: Ectomancer.ObanBridge.Queries
+    @doc false
+    defdelegate repo_delete_all(query), to: Ectomancer.ObanBridge.Queries
   end
 else
   defmodule Ectomancer.ObanBridge do
